PR Polishing - Removal of unnecessary set-define-property.ts - Refactor logging-related types
