+++ packages/airnode-deployer/jest.config.js
@@ -3,6 +3,5 @@ const config = require('../../jest.config.base');
 module.exports = {
   ...config,
   // Add custom settings below
-  setupFiles: ['<rootDir>/test/setup/init/set-define-property.ts'],
   testMatch: ['**/__tests__/**/*.[jt]s?(x)', '**/?(*.)+(spec|test).[tj]s?(x)'],
 };
