+++ packages/airnode-deployer/src/handlers/gcp/index.ts
@@ -43,6 +43,7 @@ async function initializeProvider(payload: InitializeProviderPayload, res: Respo
   const [err, initializedState] = await go(() => handlers.initializeProvider(stateWithConfig));
   if (err || !initializedState) {
     const msg = `Failed to initialize provider: ${stateWithConfig.settings.name}`;
+    logger.log(err!.toString());
     const errorLog = logger.pend('ERROR', msg, err);
     const body = { ok: false, errorLog };
     res.status(500).send(body);
