+++ packages/airnode-examples/jest-unit.config.js
@@ -5,6 +5,5 @@ module.exports = {
   // Add custom settings below
   displayName: 'unit',
   name: 'unit',
-  setupFiles: ['<rootDir>/test/setup/init/set-define-property.ts'],
   testMatch: ['**/__tests__/**/*.[jt]s?(x)', '**/?(*.)+(spec|test).[tj]s?(x)'],
 };
