+++ packages/airnode-node/src/handlers/start-coordinator.ts
@@ -1,6 +1,6 @@
 import flatMap from 'lodash/flatMap';
 import keyBy from 'lodash/keyBy';
-import { logger, go, formatDateTime } from '@api3/airnode-utilities';
+import { logger, go, formatDateTime, buildBaseOptions } from '@api3/airnode-utilities';
 import * as calls from '../coordinator/calls';
 import * as providers from '../providers';
 import { reportHeartbeat } from '../reporting';
@@ -14,7 +14,7 @@ export async function startCoordinator(config: Config) {
   const completedAt = new Date();
 
   const durationMs = Math.abs(completedAt.getTime() - startedAt.getTime());
-  const logOptions = logger.buildBaseOptions(config, { coordinatorId: endState.coordinatorId });
+  const logOptions = buildBaseOptions(config, { coordinatorId: endState.coordinatorId });
   logger.info(`Coordinator completed at ${formatDateTime(completedAt)}. Total time: ${durationMs}ms`, logOptions);
 
   // Heartbeat is not core part of coordinator because it may return early in case there are no actionable requests
@@ -27,7 +27,7 @@ export async function startCoordinator(config: Config) {
 function createInitialCoordinatorState(config: Config) {
   const state = coordinatorState.create(config);
   const { coordinatorId } = state;
-  const logOptions = logger.buildBaseOptions(config, { coordinatorId });
+  const logOptions = buildBaseOptions(config, { coordinatorId });
 
   logger.info(`Created initial coordinator state`, logOptions);
   return state;
@@ -45,7 +45,7 @@ function getWorkerOptions(state: CoordinatorState): WorkerOptions {
 
 async function initializeProviders(state: CoordinatorState) {
   const { coordinatorId, config } = state;
-  const logOptions = logger.buildBaseOptions(config, { coordinatorId });
+  const logOptions = buildBaseOptions(config, { coordinatorId });
 
   logger.info('Forking to initialize providers', logOptions);
   const [logs, providerStates] = await providers.initialize(coordinatorId, config, getWorkerOptions(state));
@@ -67,7 +67,7 @@ function hasCoordinatorNoActionableRequests(state: CoordinatorState) {
 
 function aggregateApiCalls(state: CoordinatorState) {
   const { providerStates, config, coordinatorId } = state;
-  const logOptions = logger.buildBaseOptions(config, { coordinatorId });
+  const logOptions = buildBaseOptions(config, { coordinatorId });
 
   const flatApiCalls = flatMap(providerStates.evm, (provider) => provider.requests.apiCalls);
   const aggregatedApiCallsById = calls.aggregate(config, flatApiCalls);
@@ -78,7 +78,7 @@ function aggregateApiCalls(state: CoordinatorState) {
 
 async function executeApiCalls(state: CoordinatorState) {
   const { aggregatedApiCallsById, config, coordinatorId } = state;
-  const logOptions = logger.buildBaseOptions(config, { coordinatorId });
+  const logOptions = buildBaseOptions(config, { coordinatorId });
 
   const aggregatedApiCalls = Object.values(aggregatedApiCallsById);
   const [logs, processedAggregatedApiCalls] = await calls.callApis(
@@ -95,7 +95,7 @@ async function executeApiCalls(state: CoordinatorState) {
 
 function disaggregateApiCalls(state: CoordinatorState) {
   const { config, coordinatorId } = state;
-  const logOptions = logger.buildBaseOptions(config, { coordinatorId });
+  const logOptions = buildBaseOptions(config, { coordinatorId });
 
   // TODO: Disaggregation should be chain agnostic - currently only updated EVM providers are returned
   const [logs, evmProvidersWithApiResponses] = calls.disaggregate(state);
@@ -107,7 +107,7 @@ function disaggregateApiCalls(state: CoordinatorState) {
 
 async function initiateTransactions(state: CoordinatorState) {
   const { providerStates, config, coordinatorId } = state;
-  const logOptions = logger.buildBaseOptions(config, { coordinatorId });
+  const logOptions = buildBaseOptions(config, { coordinatorId });
 
   const sponsorProviderStates = providers.splitStatesBySponsorAddress(providerStates);
   sponsorProviderStates.forEach((sponsorProviderState) => {
@@ -128,7 +128,7 @@ async function initiateTransactions(state: CoordinatorState) {
 
 function applyChainRequestLimits(state: CoordinatorState) {
   const { config, coordinatorId, providerStates } = state;
-  const logOptions = logger.buildBaseOptions(config, { coordinatorId });
+  const logOptions = buildBaseOptions(config, { coordinatorId });
 
   logger.info('Applying chain request limits', logOptions);
   const [logs, processedProviders] = calls.applyChainLimits(config, providerStates);
@@ -154,7 +154,7 @@ async function coordinator(config: Config): Promise<CoordinatorState> {
   // =================================================================
   if (hasCoordinatorNoActionableRequests(state)) {
     const { coordinatorId } = state;
-    const logOptions = logger.buildBaseOptions(config, { coordinatorId });
+    const logOptions = buildBaseOptions(config, { coordinatorId });
     logger.info('No actionable requests detected. Returning', logOptions);
     return state;
   }
