+++ packages/airnode-utilities/jest-unit.config.js
@@ -5,6 +5,5 @@ module.exports = {
   // Add custom settings below
   displayName: 'unit',
   name: 'unit',
-  setupFiles: ['<rootDir>/test/setup/init/set-define-property.ts'],
   testMatch: ['**/?(*.)+(spec|test).[tj]s?(x)'],
 };
